OSIS-4730 split up down service into two files
